float to int

+++ Assets/Scripts/UIScript/UIGameHandler.cs
@@ -12,14 +12,14 @@ public class UIGameHandler : MonoBehaviour
     [Header("Battery UI")]
     [SerializeField] private Image batteryUI;
     [SerializeField] private TextMeshProUGUI batteryText;
-    private float batteryAmount;
-    private float maxBatteryAmount;
+    private int batteryAmount;
+    private int maxBatteryAmount;
 
     [Header("Stamina UI")]
     [SerializeField] private Image staminaUI;
     [SerializeField] private TextMeshProUGUI staminaText;
-    private float staminaAmount;
-    private float maxStaminaAmount;
+    private int staminaAmount;
+    private int maxStaminaAmount;
 
     [Header("Artifact UI")]
     [SerializeField] private TextMeshProUGUI artifactText;
@@ -129,18 +129,18 @@ public class UIGameHandler : MonoBehaviour
 
     public void SetBatteryUI(float batteryAmount, float maxBatteryAmount)
     {
-        this.batteryAmount = batteryAmount;
-        this.maxBatteryAmount = maxBatteryAmount;
+        this.batteryAmount = (int)batteryAmount;
+        this.maxBatteryAmount = (int)maxBatteryAmount;
         batteryText.text = batteryAmount.ToString() + " / " + maxBatteryAmount.ToString();
         batteryUI.fillAmount = (float)batteryAmount / maxBatteryAmount;
     }
 
     public void SetStaminaUI(float staminaAmount, float maxStaminaAmount)
     {
-        this.staminaAmount = staminaAmount;
-        this.maxStaminaAmount = maxStaminaAmount;
+        this.staminaAmount = (int)staminaAmount;
+        this.maxStaminaAmount = (int)maxStaminaAmount;
         staminaText.text = staminaAmount.ToString() + " / " + maxStaminaAmount.ToString();
-        staminaUI.fillAmount = staminaAmount / maxStaminaAmount;
+        staminaUI.fillAmount = (float)staminaAmount / maxStaminaAmount;
     }
 
     public void SetArtifactUI(int collectedArtifacts, int totalArtifactsRequired)
